some code was refactoring
