--- srcs/front/components/formanage/DefaultManage.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-const DefaultManage = props => {
-	return (
-		<div>
-			DefaultManage
-		</div>
-	);
-};
-
-DefaultManage.propTypes = {
-
-};
-
-export default DefaultManage;
