+++ srcs/front/containers/ManageMenu.js
@@ -7,7 +7,6 @@ import {
 	Container,
 	Aside,
 	ContentContainer,
-	ContentHead,
 	EditBox,
 	SideTitle,
 	SideCate,
