+++ srcs/front/css/main.css
@@ -784,6 +784,11 @@ button {
 	background: transparent;
 }
 
+.Box .errorLoginMessage {
+	color: red;
+	font-size: 16px;
+}
+
 .Box .InputBox label{
 	position: absolute;
 	top: 0;
