+++ srcs/front/hooks/useSetInput.js
@@ -1,0 +1,9 @@
+import {useState, useCallback} from 'react';
+
+export default (initValue = null) => {
+	const [value, setter] = useState(initValue);
+	const handler = useCallback((e) => {
+	  setter(e.target.value);
+	}, []);
+	return [value, setter, handler];
+  };
